Add TC.CloudGames.Api.Tests project for unit testing
This commit introduces a new test project named "TC.CloudGames.Api.Tests" to the solution. The project is configured with the .NET SDK targeting .NET 9.0 and includes essential package references for testing frameworks such as xUnit, NSubstitute, and Shouldly. Additionally, it references the main API project to facilitate testing of the API code.